final commit 6 task

+++ src/main/java/com/example/springrest/UserRestController.java
@@ -2,35 +2,39 @@ package com.example.springrest;
 
 import com.example.springrest.entity.User;
 import com.example.springrest.exceptions.UserNotFoundException;
+import com.example.springrest.exceptions.UserNotRightsException;
 import com.example.springrest.service.RepositoryStubService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import java.security.Principal;
 import java.util.List;
 
 
 //1
 @RestController
-@RequestMapping("/users")
+@RequestMapping
 public class UserRestController {
 
-
+    Principal principal;
     @Autowired
     private RepositoryStubService repositoryService;
 
-   // 2
-   // если сделать так же как и в 3, то непонятно как он будет различать
+    // 2
+    // если сделать так же как и в 3, то непонятно как он будет различать
     // либо надо когда по другому переделывать, как в видео
     // передали параметр - все, ид - только определнный
     // но тогда смысл эксепшна в getUserById
-    @GetMapping("/all")
+    @GetMapping("/users/all")
     public List<User> listAllUsers() {
         return repositoryService.findAll();
     }
 
     //3
-    @GetMapping
-    public User getUserById(/*4*/@RequestParam(name="id") Long id) {
+    @GetMapping("/users")
+    public User getUserById(/*4*/@RequestParam(name = "id") Long id) {
         User user = repositoryService.findUserById(id);
         if (user == null) {
             throw new UserNotFoundException(id);
@@ -39,15 +43,17 @@ public class UserRestController {
     }
 
     //5
-    @PostMapping
-    public User createUser(/* 6 */@RequestBody User newUser) {
+    @PostMapping("/users")
+    public User createUser(/* 6 */@RequestBody User newUser, HttpServletRequest request) {
+        if (request.isUserInRole("ROLE2") == false) throw new UserNotRightsException();
         return repositoryService.saveUser(newUser);
     }
 
     //7
-    @PutMapping
-    public User updateUser(/* 8 */@RequestParam(name="id") Long id,
-                           /* 9 */@RequestBody User updatedUser) {
+    @PutMapping("/users")
+    public User updateUser(/* 8 */@RequestParam(name = "id") Long id,
+            /* 9 */@RequestBody User updatedUser, HttpServletRequest request) {
+        if (request.isUserInRole("ROLE2") == false) throw new UserNotRightsException();
         User userToUpdate = repositoryService.findUserById(id);
         if (userToUpdate != null) {
             userToUpdate.setFirstName(updatedUser.getFirstName());
@@ -61,11 +67,15 @@ public class UserRestController {
         }
     }
 
-   // 10
-    @DeleteMapping
-    public void deleteUser(/* 11 */ @RequestParam(name="id")Long id) {
+    // 10
+    @DeleteMapping("/users")
+    public void deleteUser(/* 11 */ @RequestParam(name = "id") Long id, HttpServletRequest request) {
+        if (request.isUserInRole("ROLE2") == false) throw new UserNotRightsException();
         repositoryService.deleteById(id);
     }
 
-
+    @GetMapping("/status")
+    public String getStatus() {
+        return "Service is running";
+    }
 }
